Move Validator to the package level.
